Adding new APIs

+++ src/dlio_profiler/utils/posix_internal.cpp
@@ -2,7 +2,7 @@
 // Created by haridev on 8/23/23.
 //
 
-#include "posix_internal.h"
+#include <dlio_profiler/utils/posix_internal.h>
 
 int dlp_open(const char *pathname, int flags, ...) {
   mode_t mode;
@@ -12,8 +12,7 @@ int dlp_open(const char *pathname, int flags, ...) {
   va_start(args, flags);
   if (flags & O_CREAT) {
     mode = va_arg(args, mode_t);
-  }
-  else {
+  } else {
     mode = 0;
   }
   va_end(args);
@@ -25,14 +24,14 @@ int dlp_open(const char *pathname, int flags, ...) {
 
   if (result >= 0)
     return (int) result;
-  return -1;
+  return -1; // GCOV_EXCL_LINE
 }
 
 ssize_t dlp_write(int fd, const void *buf, size_t count) {
   return syscall(SYS_write, fd, buf, count);
 }
 
-ssize_t dlp_read(int fd, void *buf, size_t count) {
+off_t dlp_read(int fd, void *buf, size_t count) {
   return syscall(SYS_read, fd, buf, count);
 }
 
@@ -40,9 +39,9 @@ int dlp_close(int fd) {
   return syscall(SYS_close, fd);
 }
 
-int dlp_fsync(int fd) {
+int dlp_fsync(int fd) { // GCOV_EXCL_START
   return syscall(SYS_fsync, fd);
-}
+} // GCOV_EXCL_STOP
 
 ssize_t dlp_readlink(const char *path, char *buf, size_t bufsize) {
   return syscall(SYS_readlink, path, buf, bufsize);
